+++ CHANGELOG.md
@@ -16,9 +16,6 @@ to [Semantic Versioning][semver].
 - Return exit codes after time command subshell measures usage
 - Output flake build as shorter etime instead of project name
 - Include manual pages when installing using flake derivation
-
-### Docs
-
 - Correct spellings of documentation to match common wordings
 
 ### Maintenance
@@ -29,6 +26,7 @@ to [Semantic Versioning][semver].
 - Reduce the set of permission required for the action runners
 - Set action steps to a pinned version with escaped variables
 - Remove flake utilities not needed for packaging this project
+- Follow the unstable packages of nix with development inputs
 
 ## [1.1.2] - 2025-03-17